requirements.txt
aiogram==3.0.0b7
python-dotenv==1.0.0
openai>=1.0.0
Pillow==9.5.0
aiohttp==3.8.5
imagehash==4.3.1
numpy==1.26.4
piexif==1.1.3
pytest==7.2.0
pytest-asyncio==0.20.2
zipfile36==0.1.3
